refctor(utils): Common method to create extent layers with more effiecient style definition
